+++ hydromt/_typing/type_def.py
@@ -95,6 +95,7 @@ Predicate = Literal[
 
 Geom = Union[gpd.GeoDataFrame, gpd.GeoSeries]
 GpdShapeGeom = Union[Geom, BaseGeometry]
+DataLike = Union[str, SourceSpecDict, Path, Dataset, DataArray]
 
 Data = Union[Dataset, DataArray]
