+++ hydromt/gis/utils.py
@@ -13,6 +13,7 @@ from typing import Optional, Tuple
 import geopandas as gpd
 import numpy as np
 import xarray as xr
+from geopandas import GeoDataFrame
 from pyflwdir import gis_utils as gis
 from pyogrio import read_info
 from pyproj import CRS
@@ -23,6 +24,8 @@ from shapely.geometry.base import BaseGeometry
 
 from hydromt import _compat
 from hydromt._typing import Bbox, Geom, GpdShapeGeom
+from hydromt._typing.type_def import Predicate
+from hydromt.models._region.region import Region
 
 __all__ = ["spread2d", "nearest", "nearest_merge"]
 
@@ -180,76 +183,6 @@ def nearest(
         gdf2_nearest = gdf2_nearest.to_crs(3857)
     dst = gdf2_nearest.distance(pnts, align=False).values
     return gdf2.index.values[idx], dst
-
-
-def filter_gdf(gdf, geom=None, bbox=None, crs=None, predicate="intersects"):
-    """Filter GeoDataFrame geometries based on geometry mask or bounding box."""
-    gtypes = (gpd.GeoDataFrame, gpd.GeoSeries, BaseGeometry)
-    if bbox is not None and geom is None:
-        if crs is None:
-            crs = gdf.crs
-        geom = gpd.GeoSeries([box(*bbox)], crs=crs)
-    elif geom is not None and not isinstance(geom, gtypes):
-        raise ValueError(
-            f"Unknown geometry mask type {type(geom).__name__}. "
-            "Provide geopandas GeoDataFrame, GeoSeries or shapely geometry."
-        )
-    elif bbox is None and geom is None:
-        raise ValueError("Either geom or bbox is required.")
-    if not isinstance(geom, BaseGeometry):
-        # reproject
-        if geom.crs is None and gdf.crs is not None:
-            geom = geom.set_crs(gdf.crs)
-        elif gdf.crs is not None and geom.crs != gdf.crs:
-            geom = geom.to_crs(gdf.crs)
-        # convert geopandas to geometry
-        geom = geom.unary_union
-    idx = np.sort(gdf.sindex.query(geom, predicate=predicate))
-    return idx
-
-
-def parse_geom_bbox_buffer(
-    geom: Optional[Geom] = None,
-    bbox: Optional[Bbox] = None,
-    buffer: float = 0.0,
-    crs: Optional[CRS] = None,
-):
-    """Parse geom or bbox to a (buffered) geometry.
-
-    Arguments
-    ---------
-    geom : geopandas.GeoDataFrame/Series, optional
-        A geometry defining the area of interest.
-    bbox : array-like of floats, optional
-        (xmin, ymin, xmax, ymax) bounding box of area of interest
-        (in WGS84 coordinates).
-    buffer : float, optional
-        Buffer around the `bbox` or `geom` area of interest in meters. By default 0.
-    crs: pyproj.CRS, optional
-        projection of the bbox or geometry. If the geometry already has a crs, this
-        argument is ignored. Defaults to EPSG:4236.
-
-    Returns
-    -------
-    geom: geometry
-        the actual geometry
-    """
-    if crs is None:
-        crs = CRS("EPSG:4326")
-    if geom is None and bbox is not None:
-        # convert bbox to geom with crs EPGS:4326 to apply buffer later
-        geom = gpd.GeoDataFrame(geometry=[box(*bbox)], crs=crs)
-    elif geom is None:
-        raise ValueError("No geom or bbox provided.")
-    elif geom.crs is None:
-        geom = geom.set_crs(crs)
-
-    if buffer > 0:
-        # make sure geom is projected > buffer in meters!
-        if geom.crs.is_geographic:
-            geom = geom.to_crs(3857)
-        geom = geom.buffer(buffer)
-    return geom
 
 
 # REPROJ
