+++ hydromt/models/_region/region.py
@@ -10,7 +10,6 @@ from pyproj import CRS
 
 from hydromt._compat import HAS_XUGRID
 from hydromt._typing.type_def import BasinIdType
-from hydromt.data_catalog import DataCatalog
 from hydromt.models._region.specifyers import RegionSpecifyer
 
 if HAS_XUGRID:
@@ -25,7 +24,6 @@ class Region:
     def __init__(
         self,
         region: Dict[str, Any],
-        data_catalog: Optional[DataCatalog] = None,
         basin_index: Optional[Union[str, GeoDataFrame]] = None,
         hydrography: Optional[Union[str, GeoDataFrame]] = None,
         crs: Optional[CRS] = None,
@@ -33,9 +31,9 @@ class Region:
     ):
         self.logger = logger
 
+        self._data = None
         self.set(
             region=region,
-            data_catalog=data_catalog,
             hydrography=hydrography,
             basin_index=basin_index,
         )
@@ -43,7 +41,6 @@ class Region:
     def set(
         self,
         region: Dict[str, Any],
-        data_catalog: Optional[DataCatalog] = None,
         basin_index: Optional[Union[str, GeoDataFrame]] = None,
         hydrography: Optional[Union[str, GeoDataFrame]] = None,
         crs: Optional[CRS] = None,
@@ -52,16 +49,21 @@ class Region:
         self.crs = crs
         self.region_specifyer = self._parse_region(
             region=region,
-            data_catalog=data_catalog,
             hydrography=hydrography,
             basin_index=basin_index,
             logger=logger,
         )
 
+    def get_geom(self):
+        pass
+
+    @property
+    def crs(self):
+        return self
+
     @staticmethod
     def _parse_region(
         region: Dict[str, Any],
-        data_catalog: Optional[DataCatalog] = None,
         basin_index: Optional[Union[str, GeoDataFrame]] = None,
         hydrography: Optional[Union[str, GeoDataFrame]] = None,
         logger=logger,
@@ -135,3 +137,63 @@ class Region:
             raise ValueError(f"Unknown region kind: {kind}")
 
         return RegionSpecifyer(spec=flat_region_dict)  # type: ignore
+
+
+def filter_gdf(gdf: GeoDataFrame, region: Region, predicate: Predicate = "intersects"):
+    """Filter GeoDataFrame geometries based on geometry mask or bounding box."""
+
+    geom = region.get_geom()
+    if not isinstance(geom, BaseGeometry):
+        # reproject
+        if geom.crs is None and gdf.crs is not None:
+            geom = geom.set_crs(gdf.crs)
+        elif gdf.crs is not None and geom.crs != gdf.crs:
+            geom = geom.to_crs(gdf.crs)
+        # convert geopandas to geometry
+        geom = geom.unary_union
+    idx = np.sort(gdf.sindex.query(geom, predicate=predicate))
+    return idx
+
+
+def parse_geom_bbox_buffer(
+    geom: Optional[Geom] = None,
+    bbox: Optional[Bbox] = None,
+    buffer: float = 0.0,
+    crs: Optional[CRS] = None,
+):
+    """Parse geom or bbox to a (buffered) geometry.
+
+    Arguments
+    ---------
+    geom : geopandas.GeoDataFrame/Series, optional
+        A geometry defining the area of interest.
+    bbox : array-like of floats, optional
+        (xmin, ymin, xmax, ymax) bounding box of area of interest
+        (in WGS84 coordinates).
+    buffer : float, optional
+        Buffer around the `bbox` or `geom` area of interest in meters. By default 0.
+    crs: pyproj.CRS, optional
+        projection of the bbox or geometry. If the geometry already has a crs, this
+        argument is ignored. Defaults to EPSG:4236.
+
+    Returns
+    -------
+    geom: geometry
+        the actual geometry
+    """
+    if crs is None:
+        crs = CRS("EPSG:4326")
+    if geom is None and bbox is not None:
+        # convert bbox to geom with crs EPGS:4326 to apply buffer later
+        geom = gpd.GeoDataFrame(geometry=[box(*bbox)], crs=crs)
+    elif geom is None:
+        raise ValueError("No geom or bbox provided.")
+    elif geom.crs is None:
+        geom = geom.set_crs(crs)
+
+    if buffer > 0:
+        # make sure geom is projected > buffer in meters!
+        if geom.crs.is_geographic:
+            geom = geom.to_crs(3857)
+        geom = geom.buffer(buffer)
+    return geom
