+++ tests/conftest.py
@@ -1,4 +1,5 @@
 from os.path import abspath, dirname, join
+
 from pathlib import Path
 from tempfile import TemporaryDirectory
